refactor: Move NoneType to sphinx.util.typing
